changed drawText x,y,size parameter to float

+++ src/include/video/graphics_utility.h
@@ -7,6 +7,6 @@ extern void checkGLError(const char *where);
 extern void rasonly(Visual *d);
 extern void doPerspective(float fov, float ratio, float znear, float zfar);
 extern void doLookAt(float *cam, float *target, float *up);
-extern void drawText(FontTex* ftx, int x, int y, int size, const char *text);
+extern void drawText(FontTex* ftx, float x, float y, float size, const char *text);
 
 #endif
